decoupled EventMap component
