+++ src/components/EventMap.js
@@ -1,53 +1,19 @@
-import React, { useEffect, useState } from 'react';
-import { MapContainer, TileLayer, Marker, Popup, useMap } from 'react-leaflet';
-import mapMarkerSvg from '../assets/map_marker.svg';
-
-import L from 'leaflet';
-
-function MapCenterUpdater({ center, bounds }) {
-  const map = useMap();
-  useEffect(() => {
-    if (bounds) {
-      map.fitBounds(bounds, { padding: [50, 50] });
-    } 
-    else if (center) {
-      map.flyTo([center.lat, center.lon], map.getZoom());
-    }
-  }, [center, bounds, map]);
-
-  console.log(bounds);
-
-  return null;
-}
+import React, { useEffect } from 'react';
+import { MapContainer, TileLayer } from 'react-leaflet';
+import { createCustomIcon } from '../assets/icons';
+import MapCenterUpdater from './MapCenterUpdater';
+import EventMarker from './EventMarker';
+import useCurrentPosition from '../hooks/useCurrentPosition';
 
 function EventMap({ filteredEvents, center, bounds }) {
-  const [currentPosition, setCurrentPosition] = useState([29.6520, -82.3250]);  // Default to Gainesville, FL
-  const customIcon = new L.Icon({
-      iconUrl: mapMarkerSvg,
-      iconSize: [38, 95],  // Size of the icon in pixels
-      iconAnchor: [22, 94],  // Point of the icon which will correspond to marker's location
-      popupAnchor: [-3, -76],  // Point from which the popup should open relative to the iconAnchor
-  });
+  const [currentPosition, setCurrentPosition] = useCurrentPosition([29.6520, -82.3250]);  // Default to Gainesville, FL
+  const customIcon = createCustomIcon();
 
-  // Get user's current location
-  useEffect(() => {
-    navigator.geolocation.getCurrentPosition(
-      (position) => {
-        const { latitude, longitude } = position.coords;
-        setCurrentPosition([latitude, longitude]);
-      },
-      () => {
-        console.error('Could not fetch your current location.');
-      }
-    );
-  }, []);
-
-  // Reset map center
   useEffect(() => {
     if (center) {
       setCurrentPosition([center.lat, center.lon]);
     }
-  }, [center]);
+  }, [center, setCurrentPosition]);
 
   return (
     <MapContainer center={currentPosition} zoom={13} maxBounds={[[90, -180], [-90, 180]]} style={{ height: 'calc(100vh - 100px)', width: '100%' }}>
@@ -55,57 +21,9 @@ function EventMap({ filteredEvents, center, bounds }) {
         url="https://{s}.tile.openstreetmap.org/{z}/{x}/{y}.png"
         attribution='&copy; <a href="https://www.openstreetmap.org/copyright">OpenStreetMap</a> contributors'
       />
-      {/* Generate markers for each event */}
       {filteredEvents.map((event, index) => (
-        <Marker key={index} position={[event.venue.latitude, event.venue.longitude]} icon={customIcon}>
-          <Popup>
-          {event.name ? (
-            <>
-              <div>
-                <strong>{event.name}</strong> is happening at <strong>{event.venue.name}</strong> on <strong>{event.formattedDate}</strong>.
-                {event.ages && ` ${event.ages}`}
-              </div>
-              {event.artistList.length > 0 && (
-                <div>
-                  Featuring: 
-                  <ul>
-                    {event.artistList.slice(0, 3).map((artist, index) => (
-                      <li key={index}>{artist.name}</li>
-                    ))}
-                    {event.artistList.length > 3 && (
-                      <li>+ {event.artistList.length - 3} more {event.artistList.length - 3 === 1 ? 'artist' : 'artists'}</li>
-                    )}
-                  </ul>
-                </div>
-              )}
-            </>
-          ) : (
-            event.artistList.length > 0 ? (
-              <>
-                <div>
-                  <strong>{event.artistList[0].name}</strong> is playing at <strong>{event.venue.name}</strong> on <strong>{event.formattedDate}</strong>.
-                  {event.ages && ` ${event.ages}`}
-                </div>
-                {event.artistList.length > 1 && (
-                  <div>
-                    Openers include: 
-                    <ul>
-                      {event.artistList.slice(1, 3).map((artist, index) => (
-                        <li key={index}>{artist.name}</li>
-                      ))}
-                    </ul>
-                  </div>
-                )}
-              </>
-            ) : 
-            <div>Unknown event happening at <strong>{event.venue.name}</strong> on <strong>{event.formattedDate}</strong>.</div>
-          )
-          }
-          <div>Learn more <a href={event.link} target="_blank" rel="noopener noreferrer">here</a>.</div>
-          </Popup>
-        </Marker>
+        <EventMarker key={index} event={event} customIcon={customIcon} />
       ))}
-
       <MapCenterUpdater center={center} bounds={bounds} />
     </MapContainer>
   );
